Revert last. no. The raid leader may be slow or have out of date mod and then you get no icons.
i'd rather fix the problem. haven't even tested my latest fix yet. If that fix doesn't work i got another idea i'm sure will work. Just saving as a last resort as it may set the icons a little slower since it'll wait til it has all GUIDs first.

IF the sync doesn't work though, i'm going to sort the GUID in numeric/hex order in a table and apply the icons in that sequence. Similar to garajal/chimereon and similar mods. no possible conflict then. the ideal solution though is the more icon setters the better, but there has to be 0 conflict.