merge branch better-unit-event-handling
all UNIT_* events can now be registered with associated unit ids as "arguments", for example:

mod:RegisterEventsInCombat("UNIT_AURA boss1 player")

This will register the UNIT_AURA event for the given unit ids and call the normal UNIT_AURA handler. Note that it is currently not possible to register the same event multiple times with different IDs, you have to use one event for all unit IDs you need.

Old UNIT_* event registrations that do not use the arguments will default to the following unit ids: boss1-5, target, focus, and mouseover. These are the same IDs that the previous special-cased UNIT-events used.

Use UNIT_*_UNFILTERED events to get events for all unit ids.


Note: all UNIT_* events now restrict the unit ids per default. this was done for some events previously (with the default unit IDs mentioned above). This might be an incomaptible change for some mods that rely on UNIT_AURA for raid targets, these mods must be changed to use UNIT_AURA_UNFILTERED


Open for discussion:
The events that are passed to the mod are currently *not* guaranteed to be restricted to the given unit ids. The extra arguments are simply a hint to DBM-Core to avoid registration of unnecessary UNIT_*-event handlers. Example: DBM-Core registers UNIT_HEALTH for boss1-5, target, focus, and mouseover. If a mod now registers the pseudo-event "UNIT_HEALTH player", then its UNIT_HEALTH handler will not only be called for "player" events, but also for the ids that are already registered by DBM-Core or other mods.
However, this could be changed to check if the actual id is registered by the mod before calling the mod's event handler. This could save some if uId == "boss1" then ... checks.